[cli,core] move IndentListDumper to core config utils
